backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
segment-anything
scikit-learn

# AI/ML
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
transformers==4.35.0
torch==2.3.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4

torchvision